Ignore localtime resolution CVEs
As far as these CVEs tell there is nothing that can be done to both fix
the potential segfault and be able to resolve local time zones. The
suggested time version to update simply gates the unsafe behaviour
behind a feature flag `unsound_local_offset` which would be needed by
alerter.

So there is no win in being constantly reminded of an unsolvable problem.